refactor: standardize module internal import paths and update TypeORM index class.
